refactor: member 미들웨어 리팩토링
