+++ src/store/thunks/memberThunks.ts
@@ -3,18 +3,16 @@ import {
   fetchMemberProfile,
   updateMemberProfile,
 } from '@/api/member'
-import {
-  setProfile,
-  setMypageData,
-  updateMypageData,
-} from '../slices/memberSlice'
+import { setProfile, setMypageData } from '../slices/memberSlice'
 import { AppDispatch } from '../store'
 import { Profile } from '@/types/_shared/profile'
 
 // 프로필을 가져오고 store에 저장
 export const fetchProfileThunk = () => async (dispatch: AppDispatch) => {
+  let profile: Profile
+
   try {
-    const profile = await fetchMemberProfile()
+    profile = await fetchMemberProfile()
     dispatch(setProfile(profile))
     return profile
   } catch (err) {
@@ -38,9 +36,18 @@ export const fetchMypageDataThunk = () => async (dispatch: AppDispatch) => {
 export const updateProfileThunk =
   (profile: Profile) => async (dispatch: AppDispatch) => {
     try {
-      await updateMemberProfile(profile)
-      dispatch(updateMypageData(profile))
-      return profile
+      await updateMemberProfile(profile) // 프로필 업데이트
+      const updatedProfile = await fetchMemberProfile() // 업데이트 된 프로필 가져오기
+      dispatch(setProfile(updatedProfile)) // store에 저장
+
+      const updatedMypageData = await fetchMemberMypage() // 업데이트 된 마이페이지 데이터 가져오기
+      if (updatedMypageData) {
+        dispatch(setMypageData(updatedMypageData)) // store에 저장
+      } else {
+        throw new Error('마이페이지 데이터 가져오기 실패')
+      }
+
+      return updatedProfile
     } catch (err) {
       throw err
     }
