+++ src/types/_shared/mypageData.ts
@@ -1,0 +1,9 @@
+export type MypageData = {
+  profile_image_url: string
+  kakaoname: string
+  email: string
+  nickname: string
+  gender: string
+  age: string
+  mbti: string
+}
